# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  domain.StatusDoAgendamento:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - Pendente
    - Confirmado
    - Cancelado
    - Concluido
...
  request_agendamento.AgendamentoRequest:
    properties:
      catalogo_id:
        type: string
      cliente_id:
        type: string
      data_hora_inicio:
        example: "2025-01-03T08:00:00Z"
        type: string
      notas:
        maxLength: 500
        type: string
      prestador_id:
        type: string
    required:
    - catalogo_id
    - cliente_id
    - data_hora_inicio
    - prestador_id
    type: object
...
  response_agendamento.AgendamentoListPaginadoResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response_agendamento.AgendamentoResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  response_agendamento.AgendamentoResponse:
    properties:
      cliente:
        $ref: '#/definitions/response_agendamento.ClienteInfo'
      data_fim:
        type: string
      data_inicio:
        type: string
      id:
        type: string
      notas:
        type: string
      prestador:
        $ref: '#/definitions/response_agendamento.PrestadorInfo'
      servico:
        $ref: '#/definitions/response_agendamento.ServicoInfo'
      status:
        $ref: '#/definitions/domain.StatusDoAgendamento'
    type: object
  response_agendamento.BuscaDataResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response_agendamento.AgendamentoResponse'
        type: array
    type: object
  response_agendamento.ClienteInfo:
    properties:
      email:
        type: string
      id:
        type: string
      nome:
        type: string
      telefone:
        type: string
    type: object
  response_agendamento.PrestadorInfo:
    properties:
      ativo:
        type: boolean
      cpf:
        type: string
      email:
        type: string
      id:
        type: string
      nome:
        type: string
      telefone:
        type: string
    type: object
  response_agendamento.ServicoInfo:
    properties:
      categoria:
        type: string
      duracao:
        type: integer
      id:
        type: string
      nome:
        type: string
      preco:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    API para gestão de clientes e serviços com autenticação JWT.

    Credenciais do Administrador (para testes):
    Email: admin@admin.com
    Senha: admin123
  title: API de Agendamentos
  version: "1.0"
paths:
  /agendamentos:
    post:
      consumes:
      - application/json
      description: Realiza o agendamento de um serviço para um prestador em uma data
        e horário específicos. Requer autenticação JWT (Cliente ou Admin).
      parameters:
      - description: Dados do agendamento
        in: body
        name: agendamento
        required: true
        schema:
          $ref: '#/definitions/request_agendamento.AgendamentoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Agendamento criado com sucesso
          schema:
            $ref: '#/definitions/response_agendamento.AgendamentoResponse'
        "400":
          description: Dados inválidos ou formato de data incorreto
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Token não fornecido ou inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Acesso negado - apenas Clientes e Admin podem criar agendamentos
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Cliente, prestador ou serviço não encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Conflito de agenda (dia ou horário indisponível)
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cria um novo agendamento
      tags:
      - Agendamentos
  /agendamentos/cliente/{id}:
    get:
      consumes:
      - application/json
      description: Retorna todos os agendamentos de um cliente a partir da data especificada,
        ordenados por data/hora de início. Requer autenticação JWT (próprio cliente
        ou Admin).
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: string
      - description: 'Data de início da busca (formato: YYYY-MM-DD)'
        example: "2025-01-03"
        in: query
        name: data
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de agendamentos encontrados
          schema:
            $ref: '#/definitions/response_agendamento.BuscaDataResponse'
        "400":
          description: Dados inválidos ou formato de data incorreto
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Token não fornecido ou inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Acesso negado - você só pode ver seus próprios agendamentos
            ou ser Admin
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Cliente não encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Busca agendamentos de um cliente a partir de uma data
      tags:
      - Agendamentos
  /agendamentos/cliente/{id}/periodo:
    get:
      consumes:
      - application/json
      description: Retorna agendamentos de um cliente em um período específico com
        suporte a paginação. Requer autenticação JWT (próprio cliente ou Admin).
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: string
      - description: 'Data de início do período (formato: YYYY-MM-DD)'
        example: "2025-01-01"
        in: query
        name: data_inicio
        required: true
        type: string
      - description: 'Data de fim do período (formato: YYYY-MM-DD)'
        example: "2025-01-31"
        in: query
        name: data_fim
        required: true
        type: string
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (padrão: 10, máx: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista paginada de agendamentos
          schema:
            $ref: '#/definitions/response_agendamento.AgendamentoListPaginadoResponse'
        "400":
          description: Dados inválidos ou formato de data incorreto
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Token não fornecido ou inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Acesso negado - você só pode ver seus próprios agendamentos
            ou ser Admin
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Cliente não encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lista agendamentos de um cliente por período com paginação
      tags:
      - Agendamentos
  /agendamentos/prestador/{id}:
    get:
      consumes:
      - application/json
      description: Retorna todos os agendamentos de um prestador a partir da data
        especificada, ordenados por data/hora de início. Requer autenticação JWT (próprio
        prestador ou Admin).
      parameters:
      - description: ID do prestador
        in: path
        name: id
        required: true
        type: string
      - description: 'Data de início da busca (formato: YYYY-MM-DD)'
        example: "2025-01-03"
        in: query
        name: data
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de agendamentos encontrados
          schema:
            $ref: '#/definitions/response_agendamento.BuscaDataResponse'
        "400":
          description: Dados inválidos ou formato de data incorreto
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Token não fornecido ou inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Acesso negado - você só pode ver seus próprios agendamentos
            ou ser Admin
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: prestador não encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Busca agendamentos de um prestador a partir de uma data
      tags:
      - Agendamentos
  /agendamentos/prestador/{id}/periodo:
    get:
      consumes:
      - application/json
      description: Retorna agendamentos de um prestador em um período específico com
        suporte a paginação. Requer autenticação JWT (próprio prestador ou Admin).
      parameters:
      - description: ID do prestador
        in: path
        name: id
        required: true
        type: string
      - description: 'Data de início do período (formato: YYYY-MM-DD)'
        example: "2025-01-01"
        in: query
        name: data_inicio
        required: true
        type: string
      - description: 'Data de fim do período (formato: YYYY-MM-DD)'
        example: "2025-01-31"
        in: query
        name: data_fim
        required: true
        type: string
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (padrão: 10, máx: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista paginada de agendamentos
          schema:
            $ref: '#/definitions/response_agendamento.AgendamentoListPaginadoResponse'
        "400":
          description: Dados inválidos ou formato de data incorreto
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Token não fornecido ou inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Acesso negado - você só pode ver seus próprios agendamentos
            ou ser Admin
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Prestador não encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lista agendamentos de um prestador por período com paginação
      tags:
      - Agendamentos
...
securityDefinitions:
  BearerAuth:
    description: Digite "Bearer " seguido do token JWT recebido no login
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"