# backend/internal/adapters/http/agendamento/request_agendamento/request_cadastro.go
package request_agendamento

import (
	"meu-servico-agenda/internal/adapters/http"
	"meu-servico-agenda/internal/core/application/input"
	"time"

	"github.com/rs/xid"
)

type AgendamentoRequest struct {
	ClienteID      string `json:"cliente_id" binding:"required" swagger:"desc('ID do cliente que está solicitando o agendamento')"`
	PrestadorID    string `json:"prestador_id" binding:"required" swagger:"desc('ID do prestador que irá atender')"`
	CatalogoID     string `json:"catalogo_id" binding:"required" swagger:"desc('ID do serviço do catálogo que será agendado')"`
	DataHoraInicio string `json:"data_hora_inicio" binding:"required,datetime=2006-01-02T15:04:05Z07:00" example:"2025-01-03T08:00:00Z"`
	Notas          string `json:"notas,omitempty" binding:"omitempty,max=500" swagger:"desc('Notas ou observações do cliente sobre o agendamento')"`
}

func (ag *AgendamentoRequest) ToAgendamento() (*input.CadastrarAgendamentoInput, error) {
	// Convertendo DataHoraInicio de string para time.Time
	dataHoraInicio, err := time.Parse(time.RFC3339, ag.DataHoraInicio)
	if err != nil {
		return nil, http.ErrFormatoDataHoraInvalido
	}

	agendamento := input.CadastrarAgendamentoInput{
		ID:             xid.New().String(),
		ClienteID:      ag.ClienteID,
		PrestadorID:    ag.PrestadorID,
		CatalogoID:     ag.CatalogoID,
		DataHoraInicio: dataHoraInicio,
		Notas:          ag.Notas,
	}

	return &agendamento, nil
}